feat: create testing.py
Rename and move functions from tests/test_trees.py to new file peakoscope/testing.py.
