+++ tests/test_init.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 # This file is part of Peakoscope.
-# Copyright (C) 2021-2024  Eivind Tøstesen
+# Copyright (C) 2021-2025  Eivind Tøstesen
 # Peakoscope is licensed under GPLv3.
 
 
@@ -9,4 +9,4 @@ import peakoscope
 
 
 def test_version():
-    assert peakoscope.__version__ == "1.1.0.dev6"
+    assert peakoscope.__version__ == "1.1.0.dev7"
